Changed the json location to be the directory of the model, rather than the root directory. Fixed setter type of integer parameter. Changed focus type for buttons, to make sure that clicking on the start button, will grab the focus from the parameter controls, which trigger the setter.

+++ ViewModel/Parameters/IntegerModel.h
@@ -25,7 +25,7 @@ public:
     {
         return ParameterModel::getValue().value<int>();
     }
-    void setValue(float value)
+    void setValue(int value)
     {
         ParameterModel::setValue(QVariant::fromValue<int>(value));
     }
